requirements.txt
aiogram>=3.6
python-dotenv>=1.0.1
pydantic>=2.8